Use marker logic with requires_to_conda
